Crate fedimint_mintv2_common

Modules

config

endpoint_constants

Structs

Denomination

Compact representation of a power-of-2 amount denomination Represents 2^denomination msats

MintCommonInit

MintInputV0

MintModuleTypes

MintOutputBlindSignature

MintOutputOutcome

MintOutputV0

Note

A verifiable one time use IOU from the mint.

UnknownMintInputVariantError

UnknownMintOutputVariantError

Enums

MintConsensusItem

The mint module currently doesn’t define any consensus items and generally throws an error on encountering one. To allow old clients to still decode blocks in the future, should we decide to add consensus items, this has to be an enum with only a default variant.

MintInput

MintInputError

MintOutput

MintOutputError

RecoveryItem

Recovery item for stateless ecash recovery

Constants

KIND

MODULE_CONSENSUS_VERSION

Functions

nonce_message